pyproject.toml
[project]
name = "follower-counter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "luma-led-matrix>=1.7.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
